Refactoring internal file structure
