+++ src/cli.py
@@ -5,7 +5,7 @@ import logging
 from src.renderers import (
     certifications_to_gitbook, inventory_builder
 )
-from src.masonry import Masonry
+from masonry.certification_builder import CertificationBuilder
 from src import template_generator
 from src import utils
 
@@ -53,8 +53,8 @@ def certs(certification, data_dir, output_dir):
     utils.create_dir(output_dir)
     certs_dir = os.path.join(data_dir, 'certifications')
     if verify_certification_path(certification, certs_dir):
-        masonry = Masonry(data_dir)
-        masonry.export_certification(certification, output_dir)
+        builder = CertificationBuilder(data_dir)
+        builder.export_certification(certification, output_dir)
         click.echo('Certification created in: `{0}`'.format(output_dir))
